Modifie l'architecture pour une meilleure visibilité sur les Repos.
