+++ app/src/main/java/fr/arthur/musicplayer/helpers/AppModule.kt
@@ -2,12 +2,21 @@ package fr.arthur.musicplayer.helpers
 
 import androidx.room.Room
 import fr.arthur.musicplayer.helpers.AppConstants.DB_NAME
-import fr.arthur.musicplayer.repositories.IMusicRepository
+import fr.arthur.musicplayer.repositories.AlbumRepository
+import fr.arthur.musicplayer.repositories.ArtistRepository
 import fr.arthur.musicplayer.repositories.MusicRepository
+import fr.arthur.musicplayer.repositories.PlaylistRepository
+import fr.arthur.musicplayer.repositories.ScannerRepository
+import fr.arthur.musicplayer.repositories.interfaces.IAlbumRepository
+import fr.arthur.musicplayer.repositories.interfaces.IArtistRepository
+import fr.arthur.musicplayer.repositories.interfaces.IMusicRepository
+import fr.arthur.musicplayer.repositories.interfaces.IPlaylistRepository
+import fr.arthur.musicplayer.repositories.interfaces.IScannerRepository
 import fr.arthur.musicplayer.room.AppDatabase
 import fr.arthur.musicplayer.usecase.GetAllArtistsUseCase
 import fr.arthur.musicplayer.usecase.GetAllMusicsUseCase
 import fr.arthur.musicplayer.usecase.GetAllPlaylistUseCase
+import fr.arthur.musicplayer.usecase.ScannerUseCase
 import fr.arthur.musicplayer.viewModel.ArtistListViewModel
 import fr.arthur.musicplayer.viewModel.MusicListViewModel
 import fr.arthur.musicplayer.viewModel.PlayListListViewModel
@@ -22,7 +31,7 @@ val appModule = module {
             AppDatabase::class.java,
             DB_NAME
         )
-            .fallbackToDestructiveMigration(true)
+//            .fallbackToDestructiveMigration(true)
             .build()
     }
 
@@ -35,14 +44,24 @@ val appModule = module {
     single { FolderUriStore(get()) }
     single { MusicScanner(get(), get()) }
 
-    single { MusicRepository(get(), get(), get(), get(), get()) }
+    single { MusicRepository(get()) }
+    single { AlbumRepository(get()) }
+    single { ArtistRepository(get()) }
+    single { PlaylistRepository(get()) }
+    single { ScannerRepository(get(), get(), get(), get()) }
+
     single<IMusicRepository> { get<MusicRepository>() }
+    single<IAlbumRepository> { get<AlbumRepository>() }
+    single<IArtistRepository> { get<ArtistRepository>() }
+    single<IPlaylistRepository> { get<PlaylistRepository>() }
+    single<IScannerRepository> { get<ScannerRepository>() }
 
     factory { GetAllMusicsUseCase(get()) }
     factory { GetAllArtistsUseCase(get()) }
     factory { GetAllPlaylistUseCase(get()) }
+    factory { ScannerUseCase(get()) }
 
-    factory { MusicListViewModel(get()) }
+    factory { MusicListViewModel(get(), get()) }
     factory { ArtistListViewModel(get()) }
     factory { PlayListListViewModel(get()) }
 }
