+++ app/src/main/java/fr/arthur/musicplayer/repositories/AlbumRepository.kt
@@ -1,0 +1,22 @@
+package fr.arthur.musicplayer.repositories
+
+import fr.arthur.musicplayer.models.Album
+import fr.arthur.musicplayer.repositories.interfaces.IAlbumRepository
+import fr.arthur.musicplayer.room.DAO.AlbumDAO
+
+class AlbumRepository(
+    private val albumDao: AlbumDAO
+) : IAlbumRepository {
+
+    override suspend fun getAllAlbums(): List<Album> {
+        return albumDao.getAll().map {
+            Album(id = it.id, name = it.name, artistId = it.artistId)
+        }
+    }
+
+    override suspend fun getAlbumsFromArtist(artistId: String): List<Album> {
+        return albumDao.getFromArtist(artistId).map {
+            Album(id = it.id, name = it.name, artistId = it.artistId)
+        }
+    }
+}
