+++ app/src/main/java/fr/arthur/musicplayer/repositories/ArtistRepository.kt
@@ -1,0 +1,16 @@
+package fr.arthur.musicplayer.repositories
+
+import fr.arthur.musicplayer.models.Artist
+import fr.arthur.musicplayer.repositories.interfaces.IArtistRepository
+import fr.arthur.musicplayer.room.DAO.ArtistDAO
+
+class ArtistRepository(
+    private val artistDao: ArtistDAO
+) : IArtistRepository {
+
+    override suspend fun getAllArtists(): List<Artist> {
+        return artistDao.getAll().map {
+            Artist(id = it.id, name = it.name)
+        }
+    }
+}
