--- app/src/main/java/fr/arthur/musicplayer/repositories/IMusicRepository.kt
@@ -1,8 +1,0 @@
-package fr.arthur.musicplayer.repositories
-
-import fr.arthur.musicplayer.models.Artist
-import fr.arthur.musicplayer.models.Playlist
-
-interface IMusicRepository {
-
-}
