+++ app/src/main/java/fr/arthur/musicplayer/repositories/PlaylistRepository.kt
@@ -1,0 +1,16 @@
+package fr.arthur.musicplayer.repositories
+
+import fr.arthur.musicplayer.models.Playlist
+import fr.arthur.musicplayer.repositories.interfaces.IPlaylistRepository
+import fr.arthur.musicplayer.room.DAO.PlaylistDAO
+
+class PlaylistRepository(
+    private val playlistDao: PlaylistDAO
+) : IPlaylistRepository {
+
+    override suspend fun getAllPlayLists(): List<Playlist> {
+        return playlistDao.getAll().map {
+            Playlist(id = it.id, name = it.name, numberOfMusics = it.numberOfMusics)
+        }
+    }
+}
