+++ app/src/main/java/fr/arthur/musicplayer/repositories/interfaces/IArtistRepository.kt
@@ -1,0 +1,7 @@
+package fr.arthur.musicplayer.repositories.interfaces
+
+import fr.arthur.musicplayer.models.Artist
+
+interface IArtistRepository {
+    suspend fun getAllArtists(): List<Artist>
+}
