+++ app/src/main/java/fr/arthur/musicplayer/repositories/interfaces/IPlaylistRepository.kt
@@ -1,0 +1,7 @@
+package fr.arthur.musicplayer.repositories.interfaces
+
+import fr.arthur.musicplayer.models.Playlist
+
+interface IPlaylistRepository {
+    suspend fun getAllPlayLists(): List<Playlist>
+}
