+++ app/src/main/java/fr/arthur/musicplayer/repositories/interfaces/IScannerRepository.kt
@@ -1,0 +1,7 @@
+package fr.arthur.musicplayer.repositories.interfaces
+
+import fr.arthur.musicplayer.models.Music
+
+interface IScannerRepository {
+    suspend fun scanAndSaveEverything(onMusicFound: (Music) -> Unit, onComplete: () -> Unit)
+}
