+++ app/src/main/java/fr/arthur/musicplayer/usecase/GetAllArtistsUseCase.kt
@@ -1,9 +1,8 @@
 package fr.arthur.musicplayer.usecase
 
 import fr.arthur.musicplayer.models.Artist
-import fr.arthur.musicplayer.repositories.IMusicRepository
-import fr.arthur.musicplayer.repositories.MusicRepository
+import fr.arthur.musicplayer.repositories.interfaces.IArtistRepository
 
-class GetAllArtistsUseCase(private val repository: MusicRepository) {
+class GetAllArtistsUseCase(private val repository: IArtistRepository) {
     suspend fun execute(): List<Artist> = repository.getAllArtists()
 }
