+++ app/src/main/java/fr/arthur/musicplayer/usecase/ScannerUseCase.kt
@@ -1,0 +1,14 @@
+package fr.arthur.musicplayer.usecase
+
+import fr.arthur.musicplayer.models.Music
+import fr.arthur.musicplayer.repositories.interfaces.IScannerRepository
+
+class ScannerUseCase(private val repository: IScannerRepository) {
+
+    suspend fun scanAndSave(
+        onMusicFound: (Music) -> Unit,
+        onComplete: () -> Unit
+    ) {
+        repository.scanAndSaveEverything(onMusicFound, onComplete)
+    }
+}
