+++ app/src/main/java/fr/arthur/musicplayer/viewModel/BaseListViewModel.kt
@@ -7,8 +7,5 @@ import kotlinx.coroutines.Job
 abstract class BaseListViewModel {
     protected val job = Job()
     protected val scope = CoroutineScope(Dispatchers.Main + job)
-
-    open fun clear() {
-        job.cancel()
-    }
+    
 }
