+++ app/src/main/java/fr/arthur/musicplayer/viewModel/MusicListViewModel.kt
@@ -3,13 +3,12 @@ package fr.arthur.musicplayer.viewModel
 import fr.arthur.musicplayer.models.Music
 import fr.arthur.musicplayer.observer.SimpleObservable
 import fr.arthur.musicplayer.usecase.GetAllMusicsUseCase
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
+import fr.arthur.musicplayer.usecase.ScannerUseCase
 import kotlinx.coroutines.launch
 
 class MusicListViewModel(
-    private val getAllMusicsUseCase: GetAllMusicsUseCase
+    private val getAllMusicsUseCase: GetAllMusicsUseCase,
+    private val scannerUseCase: ScannerUseCase
 ) : BaseListViewModel() {
     val musicsObservable = SimpleObservable<List<Music>>()
     private val musics = mutableListOf<Music>()
@@ -23,13 +22,12 @@ class MusicListViewModel(
 
             if (cached.isEmpty()) {
                 // Base vide → lancer un scan complet puis sauvegarder dans Room
-                getAllMusicsUseCase.scanAndSaveMusics(
+                scannerUseCase.scanAndSave(
                     onMusicFound = { /* on ne met rien à jour ici */ },
                     onComplete = {
                         // Après scan, recharger la base Room
                         scope.launch {
-                            val updated = getAllMusicsUseCase.loadCachedMusics()
-                            musics.addAll(updated)
+                            musics.addAll(getAllMusicsUseCase.loadCachedMusics())
                             musicsObservable.post(musics.sortedBy { it.title })
                         }
                     }
@@ -40,23 +38,17 @@ class MusicListViewModel(
                 musicsObservable.post(musics.sortedBy { it.title })
 
                 // Puis lancer un scan asynchrone en fond pour mise à jour intelligente
-                getAllMusicsUseCase.scanAndSaveMusics(
+                scannerUseCase.scanAndSave(
                     onMusicFound = { /* rien */ },
                     onComplete = {
                         scope.launch {
                             musics.clear()
-                            val updated = getAllMusicsUseCase.loadCachedMusics()
-                            musics.addAll(updated)
+                            musics.addAll(getAllMusicsUseCase.loadCachedMusics())
                             musicsObservable.post(musics.sortedBy { it.title })
                         }
                     }
                 )
             }
         }
-    }
-
-
-    fun onCleared() {
-        job.cancel()
     }
 }
