+++ app/src/main/java/fr/arthur/musicplayer/views/fragments/ArtistListFragment.kt
@@ -11,7 +11,6 @@ import fr.arthur.musicplayer.R
 import fr.arthur.musicplayer.adapters.ArtistAdapter
 import fr.arthur.musicplayer.helpers.AppConstants.COLUMNS_NUMBER
 import fr.arthur.musicplayer.viewModel.ArtistListViewModel
-import kotlinx.coroutines.launch
 import org.koin.android.ext.android.inject
 
 class ArtistListFragment : Fragment() {
@@ -35,8 +34,7 @@ class ArtistListFragment : Fragment() {
             adapter.submitList(it)
         }
 
-            viewModel.loadArtists()
-
+        viewModel.loadArtists()
 
         return view
     }
